+++ app/controllers/integral/posts_controller.rb
@@ -14,6 +14,18 @@ module Integral
       @latest_post = Integral::Post.published.where(locale: I18n.locale).order('published_at DESC').first&.decorate
       @posts = Integral::Post.published.where(locale: I18n.locale).includes(:image, :user).order('published_at DESC').paginate(page: params[:page])
       @posts = @posts.where.not(id: @latest_post.id) if @latest_post
+
+      page_title = t('integral.posts.index.title')
+      page_description = t('integral.posts.index.description')
+      if params[:page].present?
+        page_title += " - Page #{params[:page]}"
+        page_description += " - Page #{params[:page]}"
+      end
+
+      @meta_data = {
+        page_title: page_title,
+        page_description: page_description
+      }
     end
 
     # GET /<post.slug>
@@ -87,7 +99,7 @@ module Integral
 
     def validate_page_has_results
       if !params[:page].nil? && Integral::Post.published.where(locale: I18n.locale).paginate(page: params[:page]).empty?
-        raise ActionController::RoutingError, 'Invalid Page Number'
+        raise_pagination_out_of_range
       end
     end
   end