+++ app/models/ckeditor/attachment_file.rb
@@ -1,9 +1,0 @@
-# CKEditor Attachment file
-class Ckeditor::AttachmentFile < Ckeditor::Asset
-  mount_uploader :data, CkeditorAttachmentFileUploader, mount_on: :data_file_name
-
-  # Sets the URL of the thumbnail(?)
-  def url_thumb
-    @url_thumb ||= Ckeditor::Utils.filethumb(filename)
-  end
-end